fix: 게시글 확인 로직 추가
- 존재하지 않거나 이미 삭제됐음에도 200 OK를 반환하던 문제 해결을 위해 게시글 확인 로직 추가함

+++ src/main/kotlin/com/boardkt/post/service/PostService.kt
@@ -54,6 +54,9 @@ class PostService(
 
     @Transactional
     fun deletePost(id: Long) {
+        if (!postJpaRepository.existsById(id)) {
+            throw RuntimeException("존재하지 않거나 이미 삭제된 게시글입니다.")
+        }
         postJpaRepository.deleteById(id)
     }
 }
